backend scafolding completed and updates to routes,swagger
