+++ backend/config.py
@@ -11,12 +11,14 @@ class Config:
 class DevConfig(Config):
     SQLALCHEMY_DATABASE_URI="sqlite:///"+os.path.join(BASE_DIR,'dev.db')
     DEBUG=True
-    SQLALCHEMY_ECHO=True
+    # SQLALCHEMY_ECHO=True
 
 class ProdConfig(Config):
+    # future implementation to publish on heroku
     pass
 
 class TestConfig(Config):
+    # future implementation to build Testcases
     pass
