+++ backend/run.py
@@ -1,0 +1,6 @@
+from main import create_app
+from config import DevConfig,ProdConfig
+
+if __name__ == '__main__':
+    app=create_app(DevConfig)
+    app.run()
